GenericDialogPlus is not used
